pwmss2.dtsi: use absolute addresses
This allows for compatibility with the default declarations for pwmss,
hence to use uio for only a subset of the pwmss subdevices.

Note that in all cases you will need to override the 'ranges' property
of the pwmss (the default doesn't span the whole 4K) and create the
'module' subdevice.